refactor: switch over to SpringBoot project
With this commit, everything that was before is removed except the
.gitignore, .gitattributes and LICENSE files.

We shall re-add the documentation from here, as well as check if we
can switch the adr doc generation and the pre-commit handling to maven.
If we cannot switch it, we might still have to use yarn for the docs and
the pre-commit.